backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.22",
    "celery[redis]>=5.6.2",
    "chromadb>=1.4.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-aws>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=1.1.6",
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.9",
    "pinecone>=8.0.0",
    "pinecone-text>=0.11.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "rank-bm25>=0.2.2",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
